Fold folds and history into folders
The last two source files at the root of graph-core. Both are smaller than
the six already folded — 239 and 170 code lines against 651-1,095 — so the
question was whether a folder helps or just adds a hop. It helps for
different reasons in each.

folds/ (4 modules). `certainty` and `monoid` are pure algebra, `cache` is
the one stateful thing in the file, and `compute` is what walks the graph.
Those are three different kinds of code that happened to share a name, and
the memo table alone is 268 lines.

history/ (5 modules). This one declared SIX sections, and I merged two:
construction and push are one gesture and each is under thirty lines, so
splitting them would have produced files too small to navigate to. The rest
— read, clear, coalesce, scrub — stand on their own.

Both folders are value-acyclic, and each turned out to need exactly ONE
cross-module import: `monoid` reads `weakestCertainty` from `certainty`,
and `stack` reads `coalesceEntries` from `coalesce`. That the seams were
that clean is the argument for the split.

`folds.test.ts` and `history.test.ts` moved beside their subjects.

WHAT IS LEFT AT THE ROOT, deliberately: `index.ts` (the package's public
surface), the compile-only tuple fixture, and 32 cross-cutting tests. Every
source module now lives in a folder — eight of them, all value-acyclic,
with `types` carrying the one type-only cycle that `import type` erases.

Gate: 7/7 packages typecheck, noUnusedLocals clean, lint:packages and app
lint 0 errors, app tsc clean, 1555 unit tests across the same 85 files.
